+++ docs/changelog.md
@@ -19,6 +19,13 @@ Added:
 
 - A helper function named [fit_gaussian()][extra.utils.fit_gaussian] (!131).
 
+Fixed:
+
+- Sometimes an XGM will record the wrong number of pulses in its slow data
+  property, which would cause the pulse energy to not be retrieved properly. Now
+  the XGM will rely only on the fast data to find the number of pulses in
+  `.pulse_energy()` (!153).
+
 ## [2024.1.1]
 
 Added: